requirements.txt
nextcord
tortoise-orm
aerich